vision/requirements.txt
mss>=9.0
opencv-python-headless>=4.9
numpy>=1.26
Pillow>=10.0
imagehash>=4.3
ocrmac>=0.5
